// jungsi-frontend/src/stores/server/features/explore/early-subject/interfaces.ts
export interface IExploreEarlySubjectStep4Item {
  id: number;
  name: string;
  recruitment_number: number | null;
  university: {
    id: number;
    name: string;
    region: string;
    code: string;
    establishment_type: string;
  };
  admission: {
    id: number;
    name: string;
    year: number;
    basic_type: "일반" | "특별";
  };
  general_field: {
    id: number;
    name: string;
  };
  scores: {
    grade_50_cut: number | null;
    grade_70_cut: number | null;
    convert_50_cut: number | null;
    convert_70_cut: number | null;
    risk_plus_5: number | null;
    risk_plus_4: number | null;
    risk_plus_3: number | null;
    risk_plus_2: number | null;
    risk_plus_1: number | null;
    risk_minus_1: number | null;
    risk_minus_2: number | null;
    risk_minus_3: number | null;
    risk_minus_4: number | null;
    risk_minus_5: number | null;
  } | null;
}